// src/app/Shared/Enums/FixedPriceProjectEnum.ts
export enum Currency {
    USD = 'USD',
    EUR = 'EUR',
    GBP = 'GBP',
    JPY = 'JPY',
    AUD = 'AUD',
    CAD = 'CAD',
    CHF = 'CHF',
    CNY = 'CNY',
    SEK = 'SEK',
    NZD = 'NZD'
  }
  
  export enum ExperienceLevel {
    Entry = 'Entry',
    Intermediate = 'Intermediate',
    Expert = 'Expert'
  }